engine/requirements.txt
numpy
ifcopenshell
PyQt5
PyOpenGL 